Revert "chg: p_Shift: int->long"
This reverts commit 7e0dcb975494701bda4e5a9bc78db2b7518472d5.

+++ libpolys/polys/monomials/p_polys.cc
@@ -4703,10 +4703,10 @@ int p_LowVar (poly p, const ring r)
 /*2
 * verschiebt die Indizees der Modulerzeugenden um i
 */
-void p_Shift (poly * p,long i, const ring r)
+void p_Shift (poly * p,int i, const ring r)
 {
   poly qp1 = *p,qp2 = *p;/*working pointers*/
-  long     j = p_MaxComp(*p,r),k = p_MinComp(*p,r);
+  int     j = p_MaxComp(*p,r),k = p_MinComp(*p,r);
 
   if (j+i < 0) return ;
   BOOLEAN toPoly= ((j == -i) && (j == k));
