+++ libpolys/polys/monomials/p_polys.h
@@ -2105,7 +2105,7 @@ int   p_LowVar (poly p, const ring r);
 
 /*----------------------------------------------------*/
 /// shifts components of the vector p by i
-void p_Shift (poly * p,long i, const ring r);
+void p_Shift (poly * p,int i, const ring r);
 /*----------------------------------------------------*/
 
 int p_Compare(const poly a, const poly b, const ring R);
